requirements.txt
web3==6.15.1
solana==0.34.2
solders==0.21.0
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
aiohttp==3.9.3
requests==2.31.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1
hiredis==2.3.2

# Monitoring
prometheus-client==0.19.0

# WebSocket support
websockets==12.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background tasks
celery==5.3.6
